changed Note model to use enum for types instead of booleans
reasoning: more flexible for the future since can change types without database migrations

ATTENTION: instead of creating a database migration, this is done by redoing the initial migration.
This means that current tables are meant to be completely deleted and recreated.
The reasoning is that we are new in the development process and still have no real data.
To pull in this change do:

$ cd backend
$ rm app.db
$ flask db upgrade